Refactored to make the fields more flexible.
Before this, the YAML file could only support five specific common fields at the top of the file.  The ones that are applied to all issues.  Now you can specify any arbitrary fields and it will use them.  As long as JIRA is happy with the fields you use, it will work.